requirements.txt
requests>=2.32
pandas>=2.2
beautifulsoup4>=4.12
python-dateutil>=2.9
unidecode>=1.3
pyarrow>=18.0
